refactor: deno_http op crate (#11335)
