+++ Cargo.toml
@@ -12,6 +12,7 @@ members = [
   "extensions/console",
   "extensions/crypto",
   "extensions/fetch",
+  "extensions/http",
   "extensions/net",
   "extensions/timers",
   "extensions/url",
@@ -55,6 +56,8 @@ opt-level = 3
 opt-level = 3
 [profile.bench.package.deno_runtime]
 opt-level = 3
+[profile.bench.package.deno_http]
+opt-level = 3
 [profile.bench.package.deno_net]
 opt-level = 3
 [profile.bench.package.rusty_v8]
@@ -80,6 +83,8 @@ opt-level = 3
 [profile.release.package.deno_core]
 opt-level = 3
 [profile.release.package.deno_runtime]
+opt-level = 3
+[profile.release.package.deno_http]
 opt-level = 3
 [profile.release.package.deno_net]
 opt-level = 3
