+++ cli/build.rs
@@ -71,6 +71,8 @@ fn create_compiler_snapshot(
   op_crate_libs.insert("deno.net", deno_net::get_declaration());
   op_crate_libs
     .insert("deno.net_unstable", deno_net::get_unstable_declaration());
+  op_crate_libs
+    .insert("deno.http_unstable", deno_http::get_unstable_declaration());
 
   // ensure we invalidate the build properly.
   for (_, path) in op_crate_libs.iter() {
@@ -312,6 +314,10 @@ fn main() {
   println!(
     "cargo:rustc-env=DENO_NET_UNSTABLE_LIB_PATH={}",
     deno_net::get_unstable_declaration().display()
+  );
+  println!(
+    "cargo:rustc-env=DENO_HTTP_UNSTABLE_LIB_PATH={}",
+    deno_http::get_unstable_declaration().display()
   );
 
   println!("cargo:rustc-env=TARGET={}", env::var("TARGET").unwrap());
