+++ runtime/ops/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod fs;
 pub mod fs_events;
+pub mod http;
 pub mod io;
 pub mod os;
 pub mod permissions;
